refactor: extract IntlProvider and lang switcher logic into own component
